Feat: Updates localization for German
Updates localization files for the German language, including both formal and informal variants.

This change ensures that all text within the plugin is properly translated and displayed to German-speaking users. It includes updates to the .po and .mo files, as well as JSON files for block editor localization.
The change also fixes the renaming of the formal language files.